renamed executable to bmc
