+++ client/.gitignore
@@ -1,0 +1,3 @@
+*.json
+*.swp
+bmc
